Git & GitHub Assignment 3
Git and Github part 2 
Assignment 3


(1)	 How to check if git is available on your system?
 If we want to check the git is available or not on our system then we can use the      command:
“ git --version “ in the command line or terminal. If git is installed, the command will return the version number of git installed on your system. If git is not installed, the command will return an error message indicating that git is not recognized as a command.
(2)	 How to initialize a new git repository?
We can initialize a new git repository with the help of command :
“Git init”  
The command “ git init “creates a new subdirectory named .git that contains all of your necessary repository files.
Then you can start adding files and making commits to the repository.

(3)	 How to tell git about name and email?    
We can tell to git about our name and email with the help of the command : 
git config –global user.name “vinaykumar860964”
git config –global user.email “vinaykumar@860964gmail.com”

(4)	 How to add a file to the staging area?
To add a file to the staging area in Git, we can use the command “git add” followed by the file name or path. For example, to add a file named "file.txt" in the current directory, you would run the following command:
•	git add file.txt
•	git add .
•	git add -pw
•	git add "*.txt"

After adding the file in the staging area we can use the command “git commit” to save the change of the repository.

(5)	  How to remove a file from the staging area?  
If we want to remove the file from the staging area then we can use the command “git reset” followed by file name or path.
Lets see one example  to remove a file named from the staging area "file.txt" we can run the following command:
•	git reset file.txt


•	git reset


•	git reset HEAD

It will unstage all the files that were added to the staging area but it will not remove any files from the working directory.

(6)	 How to make a commit ?
If we want to make a commit in Git, we  will first need to add our changes to the staging area using the “ git add “ command. Once the changes are in the staging area, you can use the “ git commit “ command to save the changes to the repository.
Lets try to understand with an example how can we make a commit: 
•	Make changes to one or more files in our working directory.
•	We can use “ git add “ to add the changed files to the staging area.
•	Use the “git commit” command to save the changes to the repository.
•	git commit -m "we can write here our commit message"
     Once we can  run the ” git commit “ command, our changes will be saved to the repository and a new commit will be created with a unique hash.

(7)	 How to send your changes to a remote repository?
If we want to send our changes to a remote repository in Git, we will first need to make a commit of your local changes using the git commit command. Once we have made a commit, we can use the git push command to send your changes to the remote repository.

Lets see  an example of how to send our  changes to a remote repository:


•	We can Make changes to one or more files in our working directory.


•	We can  git  add to add the changed files to the staging area.


•	We can Use the git commit command to save the changes to the local repository.


•	We can Use the git push command to send the changes to the remote repository.

          git push <remote> <branch>

<branch> is the name of the branch we want to push our changes.

For example, if we want to push your changes to the master branch of the remote repository named "origin", we would run the following command:
git push origin master
Before pushing our  changes make sure we  are in the correct branch and our local repository is up-to-date with the remote repository by pulling the latest changes using the git pull command.

(8)	 What is the difference between clone and pull?
As I study with pw skills “git clone” and “ git pull “  are both used to copy a repository from a remote source to a local machine, but they are used for different purposes.


•	git clone is used to create a copy of a remote repository on a local machine for the first time. When we run “git clone” , it creates a new directory with the same name as the repository and copies all the files and history from the remote repository to the local machine.
          git clone <repository-url>
•	git pull, on the other hand, is used to update an existing local repository with changes from a remote repository. When we run git pull, it fetches the changes from the remote repository and merges them into our local repository. The command can be used to synchronize our local repository with the remote repository, by pulling in new commits and updating the local branches.
         git pull <remote> <branch>
                  <remote> is the name of the remote repository you want to pull from. 
                   For example, if you want to pull the changes from the master branch of          the remote repository named "origin", you would run the following command:git pull origin master.