Fundamentals of java Assignment 4.
Fundamentals of Java
Assignment 4



1.	 What is a programming language?

A programming language is like a set of instructions that are used to tell the computer what to do.

Real life Example : ordering food at a fast-food restaurant. The customer (programmer) places an order (code) by telling the cashier (compiler) what they want to eat (program). The cashier (compiler) then communicates the order (code) to the kitchen staff (computer) who prepare the food (execute the program) and give it to the customer (programmer) to eat.Just like how a customer uses a language to communicate their order to the restaurant, a programmer uses a programming language to communicate their instructions to the computer. The computer then interprets these instructions, and performs the corresponding actions.







2.	Why do we need a programming language?

We need a programming language because it serves as a means of communication between the developer and the computer. Without a programming language, it would be difficult to give the computer specific instructions and to create software and applications.
programming languages also provide a set of tools and structures that help the developer to write efficient and maintainable code, such as variables, functions, classes, and libraries. They also provide error handling and debugging tools that allow the developer to detect and fix issues in their code.
So I can say that programming languages are very needed for us .





3.	What are the features of java ?

There are following features of java : 
Object-Oriented: Java is an object-oriented programming language, which means it uses objects, classes, and interfaces to represent data and behavior.

*Platform Independent: Java code is compiled into an intermediate form called bytecode, which can be run on any platform that has a Java Virtual Machine (JVM) installed. This makes Java highly portable.

Simple: Java has a simple and easy-to-learn syntax, which makes it a great language for beginners.

Secure: Java provides a high level of security by checking bytecode for any illegal code before it is executed.

Multithreaded: Java supports multithreading, which means it can run multiple threads of execution at the same time.

Dynamic: Java supports dynamic binding, which means that objects can be linked to the methods and variables they need at runtime.

High Performance: Java code can be optimized for high performance by using the Just-In-Time (JIT) compiler.

There are some main features of java .



4.	What is an object?

Objects are instances of classes and can be created and manipulated at runtime. Objects are often used to model real-world entities and concepts, and they can be used to organize and structure code in a way that is easy to understand and maintain.





5.	 What is a class?

A class is a blueprint for creating objects ,providing initial values for state , and implementations of behavior (member functions or methods). A class can be thought of as a template for creating objects, and it defines a set of properties and methods that are common to all objects of that class.
For example, you could define a class called "Car" that has properties such as make, model, and year, and methods such as start, stop, and drive. Then you could create multiple objects that are instances of the "Car" class, each with their own specific values for the properties, but they would all have the same methods available to them.



6.	Explain about the main() method in java.

method that the Java Virtual Machine looks for and runs when a program is executed. The main method is defined as a static method with the following : 
•	public static void main(String[] args)
•	The keyword "public" means that the method can be called from anywhere, 
•	The keyword "static" means that the method can be called without creating an instance of the class, and
•	The keyword "void" means that the method does not return a value. 
•	The "String[] args" parameter is an array of strings that can be passed to the program when it is run from the command line.